updt: added the possibility to update the group title from the plot profile.
Once the plot profile is opened, in the groups card, there will be a edit button at the end of each grop's container.
Clicking that button will reveal an input field where the update title should be put. Once the check button is clicked, the new title is saved to the group entity and the visuals of the groups card is updated.