feate: making sure to validate the date of fertilizers + better toggling for fertilizer groups + fertilizer data saving
When a date is submitted for a fertilizer group, it should now follow
the already established date standards.

Excluding the date from the `toggle new group` requirements.
Making sure that the fertilizer group can have empty quantity, removing
only when the fertilizer type is set to `No fertilizer`

Making sure we update the fertilizer data properly