bugfix: make sure that we save the fertilizer shape to the db.
We need to save the shape of the FTLZ object to the databese and not
delete any not filled fertilizers, because upon rendering, there will be
a need to create the elements for the fertilizers on the fly.
Deleting unset values before setting to the database will do more harm than good.